feat: production readiness fixes (P0 + critical gaps)
- Fix manifest drift: parse Robocopy per-file failures
- Fix file_size column overflow: Integer -> BigInteger
- Wire up email notifications in on_failure hook
- Fix Rclone exit code 5 mapping (CLOUD_PARTIAL -> CLOUD_FAILED for retries)
- Add /XJ to Robocopy to exclude junction points
- Add manifest.db backup task after successful runs
- Add post-backup integrity verification (rclone check)
- Batch manifest lookups in scanner (200K+ files)
- Rewrite setup_email_notifications.py with real Prefect automations
- Add install_services.bat / uninstall_services.bat for deployment
- Add .env.example for deployment documentation
- Add UI /health endpoint for NSSM monitoring
- Update AGENTS.md, DECISIONS.md, STATE.md with current status

+++ models/manifest_model.py
@@ -1,6 +1,6 @@
 """SQLAlchemy model for the file manifest database."""
 
-from sqlalchemy import Column, Index, Integer, REAL, String, Text, create_engine, event
+from sqlalchemy import BigInteger, Column, Index, Integer, REAL, Text, create_engine, event
 from sqlalchemy.orm import DeclarativeBase, sessionmaker
 
 
@@ -13,7 +13,7 @@ class FileManifest(Base):
 
     file_id = Column(Text, primary_key=True)
     relative_path = Column(Text, nullable=False, unique=True)
-    file_size = Column(Integer, nullable=False)
+    file_size = Column(BigInteger, nullable=False)
     last_modified_timestamp = Column(REAL, nullable=False)
     checksum = Column(Text, nullable=False, default="pending")
     last_seen_at = Column(Text, nullable=False)
